+++ rngay-service-authority/src/main/java/com/rngay/service_authority/dao/UserRoleDao.java
@@ -1,7 +1,7 @@
 package com.rngay.service_authority.dao;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.rngay.feign.platform.UserRoleDTO;
+import com.rngay.feign.authority.UserRoleDTO;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
@@ -10,6 +10,6 @@ import java.util.List;
 @Repository
 public interface UserRoleDao extends BaseMapper<UserRoleDTO> {
 
-    List<UserRoleDTO> getRoleId(@Param("userId") Integer userId);
+    List<UserRoleDTO> getRoleId(@Param("userId") Long userId);
 
 }
