+++ rngay-service-authority/src/main/java/com/rngay/service_authority/model/UAUserRole.java
@@ -9,34 +9,40 @@ public class UAUserRole implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy= GenerationType.IDENTITY)
-	private Integer id;
-	private Integer userId;
-	private Integer roleId;
+	private Long id;
+	private Long userId;
+	private Long roleId;
 	private Integer checked;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+
+	public void setId(Long id) {
 		this.id = id;
 	}
-	public Integer getUserId() {
+
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(Integer userId) {
+
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
-	public Integer getRoleId() {
+
+	public Long getRoleId() {
 		return roleId;
 	}
-	public void setRoleId(Integer roleId) {
+
+	public void setRoleId(Long roleId) {
 		this.roleId = roleId;
 	}
+
 	public Integer getChecked() {
 		return checked;
 	}
+
 	public void setChecked(Integer checked) {
 		this.checked = checked;
 	}
-
 }
