+++ rngay-service-authority/src/main/java/com/rngay/service_authority/model/UAUserToken.java
@@ -10,7 +10,7 @@ public class UAUserToken implements Serializable {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private Date createTime;
 
@@ -20,13 +20,13 @@ public class UAUserToken implements Serializable {
 
     private Date updateTime;
 
-    private Integer userId;
+    private Long userId;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -62,12 +62,11 @@ public class UAUserToken implements Serializable {
         this.updateTime = updateTime;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
-
 }
