+++ rngay-service-authority/src/main/java/com/rngay/service_authority/service/impl/MenuUrlServiceImpl.java
@@ -2,9 +2,9 @@ package com.rngay.service_authority.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.rngay.feign.platform.MenuUrlDTO;
-import com.rngay.feign.platform.UpdateUrlDTO;
-import com.rngay.feign.platform.UrlDTO;
+import com.rngay.feign.authority.MenuUrlDTO;
+import com.rngay.feign.authority.UpdateUrlDTO;
+import com.rngay.feign.authority.UrlDTO;
 import com.rngay.service_authority.dao.MenuUrlDao;
 import com.rngay.service_authority.dao.UrlDao;
 import com.rngay.service_authority.service.MenuUrlService;
@@ -37,7 +37,7 @@ public class MenuUrlServiceImpl extends ServiceImpl<MenuUrlDao, MenuUrlDTO> impl
     }
 
     @Override
-    public List<MenuUrlDTO> loadUrl(Integer id) {
+    public List<MenuUrlDTO> loadUrl(Long id) {
         return menuUrlDao.selectList(new QueryWrapper<MenuUrlDTO>().eq("menu_id", id).eq("checked", 1));
     }
