+++ rngay-service-authority/src/main/java/com/rngay/service_authority/service/impl/UserRoleServiceImpl.java
@@ -2,9 +2,9 @@ package com.rngay.service_authority.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.rngay.feign.platform.OrgRoleDTO;
-import com.rngay.feign.platform.UserRoleDTO;
-import com.rngay.feign.platform.UserRoleUpdateDTO;
+import com.rngay.feign.authority.OrgRoleDTO;
+import com.rngay.feign.authority.UserRoleDTO;
+import com.rngay.feign.authority.UserRoleUpdateDTO;
 import com.rngay.service_authority.dao.OrgRoleDao;
 import com.rngay.service_authority.dao.UrlDao;
 import com.rngay.service_authority.dao.UserRoleDao;
@@ -28,7 +28,7 @@ public class UserRoleServiceImpl extends ServiceImpl<UserRoleDao, UserRoleDTO> i
     private UrlDao urlDao;
 
     @Override
-    public List<UserRoleDTO> load(Integer userId) {
+    public List<UserRoleDTO> load(Long userId) {
         return userRoleDao.selectList(new QueryWrapper<UserRoleDTO>()
                 .eq("checked", 1).eq("user_id", userId));
     }
@@ -69,7 +69,7 @@ public class UserRoleServiceImpl extends ServiceImpl<UserRoleDao, UserRoleDTO> i
     }
 
     @Override
-    public List<String> loadUrlByUserId(Integer orgId, Integer userId) {
+    public List<String> loadUrlByUserId(Integer orgId, Long userId) {
         List<UserRoleDTO> roleId = userRoleDao.getRoleId(userId);
         if (roleId == null || roleId.isEmpty()) return new ArrayList<>();
         if (orgId != null && orgId > 0) {
